requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
google-generativeai==0.4.1
torch==2.2.1
gymnasium==0.29.1
